refactor!: rename domain/task*s*_list to task_list
