+++ todoapp/application/task/commands/create_task.py
@@ -9,7 +9,7 @@ from todoapp.application.task.interfaces import TaskRepo,TaskListGetter
 from todoapp.application.task_list.exceptions import TaskListAccessError
 from todoapp.domain.common.constants import Operation
 from todoapp.domain.task.entities import Task
-from todoapp.domain.tasks_list.value_objects import ListId
+from todoapp.domain.task_list.value_objects import ListId
 from todoapp.domain.user.entities import UserId
