+++ todoapp/application/task_list/commands/share_task_list.py
@@ -6,7 +6,7 @@ from todoapp.application.common.interfaces.uow import UnitOfWork
 from todoapp.application.task_list.exceptions import TaskListAccessError
 from todoapp.application.task_list.interfaces import TaskListRepo
 from todoapp.domain.common.constants import Operation
-from todoapp.domain.tasks_list.value_objects import ListId, SharingRule
+from todoapp.domain.task_list.value_objects import ListId, SharingRule
 from todoapp.domain.user.entities import UserId
