+++ todoapp/application/task_list/queries/get_list_by_id.py
@@ -6,7 +6,7 @@ from todoapp.application.task_list import dto
 from todoapp.application.task_list.exceptions import TaskListAccessError
 from todoapp.application.task_list.interfaces import TaskListRepo,TaskInListFinder
 from todoapp.domain.common.constants import Operation
-from todoapp.domain.tasks_list.value_objects import ListId
+from todoapp.domain.task_list.value_objects import ListId
 from todoapp.domain.user.entities import UserId
