+++ todoapp/domain/task_list/entities.py
@@ -6,8 +6,8 @@ from uuid6 import uuid7
 
 from todoapp.domain.common.constants import Operation
 from todoapp.domain.common.entities import BaseEntity
-from todoapp.domain.tasks_list.exceptions import TaskListVisibilityNotModified, SharingRuleNotExistsError
-from todoapp.domain.tasks_list.value_objects import ListId, Sharing, SharingRule
+from todoapp.domain.task_list.exceptions import TaskListVisibilityNotModified, SharingRuleNotExistsError
+from todoapp.domain.task_list.value_objects import ListId, Sharing, SharingRule
 from todoapp.domain.user.entities import UserId
 
 MAX_TASK_LIST_NAME_LENGTH = 500
