+++ todoapp/presentation/api/controllers/requests/task_list.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from pydantic import BaseModel, Field
 
-from todoapp.domain.tasks_list.entities import MIN_TASK_LIST_NAME_LENGTH, MAX_TASK_LIST_NAME_LENGTH
+from todoapp.domain.task_list.entities import MIN_TASK_LIST_NAME_LENGTH, MAX_TASK_LIST_NAME_LENGTH
 
 
 class CreateTaskListRequest(BaseModel):
